pyproject.toml
[tool.poetry]
name = "parameter-mapper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
cel-python = "^0.2.0"
fastapi = "^0.115.12"
jsonpath-ng = "^1.7.0"
jinja2 = "^3.1.6"
uvicorn = "^0.34.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
